fix: move statistics function to apt page
