+++ src/routes/[library]/[section]/admin-dashboard/statistics/+page.svelte
@@ -1,17 +1,58 @@
 <script lang="ts">
 	// UI Imports
-	import { Toaster } from 'svelte-5-french-toast';
+	import toast, { Toaster } from 'svelte-5-french-toast';
 	import * as Alert from '$lib/components/ui/alert';
 	import CircleAlert from 'lucide-svelte/icons/circle-alert';
 	// Backend Imports
 	import { AdminStore } from '$lib/stores/AdminStore';
 	import { StatisticStore } from '$lib/stores/StatisticStore';
+	import { countTotalUsageLog } from '../../../../supabase/UsageLog';
+	import { page } from '$app/stores';
+        
+    const routes: Array<string> = $page.url.pathname.split('/');
+	const library: string = routes[1]; // session
+	const section: string = routes[2]; // session
 
-    export let data: { libraryName: string, librarySection: string };
-    let section:string = data.librarySection
+    export let data: { libraryName: string };
+    const librarySection:string = section
         .split('-')
         .map(word => word.charAt(0).toUpperCase() + word.slice(1))
-        .join(' ')
+        .join(' ');
+
+        async function getStatistics() {
+        // gets all statistics for the logged in admin
+        const { count, error } = await countTotalUsageLog({
+            usagelog_id: 0,
+            start: null,
+            end: null,
+            is_active: null,
+            lib_user_id: 0,
+            service_type: '',
+            library,
+            section,
+            admin_id: $AdminStore.formData.admin_id
+        });
+
+        if (error) {
+            toast.error(`Error with getting statistics: ${error}`);
+        } else if (count) {
+            $StatisticStore.total_usagelogs = count;
+
+            // const { serviceTypes, error } = await readServiceType();
+            // if (error) {
+            //     return;
+            // } else if (serviceTypes != null) {
+
+            //     $ServiceTypeStore = serviceTypes;
+            //     console.log($ServiceTypeStore)
+
+            //     // for (const service of $ServiceTypeStore) {
+            //     //     console.log(service)
+            //     // }
+            // }
+        }
+        return;
+    }
 </script>
 
 <Toaster />
@@ -20,8 +61,9 @@
 	{#if $AdminStore.formData.is_approved}
 		<div class="w-[95%] flex flex-col gap-4">
 			<h1 class="pt-10 text-3xl font-medium">Hello, {$AdminStore.formData.nickname}</h1>
-            <h2 class="text-lg text-[#636363]">Here are your statistics in {data.libraryName}, {section}</h2>
+            <h2 class="text-lg text-[#636363]">Here are your statistics in {data.libraryName}, {librarySection}</h2>
             <p>Total Usage Logs Supervised: {$StatisticStore.total_usagelogs}</p>
+            <button on:click={getStatistics}>Get Stat</button>
 		</div>
 	{:else}
 		<div class="flex h-full w-full flex-col gap-10 p-20">
